rename AbilityScoresDisplay
too wordy
